Fixed encoding error
The characeter encoding used when encoding the JSON body in
http requests did NOT match the expected encoding on the server.
I switched body encoding to encode as UTF-8 to resolve this issue.

Because most English alphabet characeters are encoded the same way
across many different encodings, this did not arise as an issue until
using latin characeters like é.